+++ clevertap-core/src/main/java/com/clevertap/android/sdk/CTXtensions.kt
@@ -2,11 +2,30 @@
 
 package com.clevertap.android.sdk
 
+import android.app.NotificationManager
 import android.content.Context
+import android.content.Context.NOTIFICATION_SERVICE
 import android.os.Build.VERSION
+import android.os.Build.VERSION_CODES
+import androidx.core.app.NotificationManagerCompat
 
 fun Context.isPackageAndOsTargetsAbove(apiLevel: Int) =
     VERSION.SDK_INT > apiLevel && targetSdkVersion > apiLevel
 
 val Context.targetSdkVersion
     get() = applicationContext.applicationInfo.targetSdkVersion
+
+fun Context.isNotificationChannelEnabled(channelId: String): Boolean =
+    if (VERSION.SDK_INT >= VERSION_CODES.O) {
+        areAppNotificationsEnabled() && try {
+            val nm = getSystemService(NOTIFICATION_SERVICE) as NotificationManager
+            nm.getNotificationChannel(channelId).importance != NotificationManager.IMPORTANCE_NONE
+        } catch (e: Exception) {
+            Logger.d("Unable to find notification channel with id = $channelId")
+            false
+        }
+    } else {
+        areAppNotificationsEnabled()
+    }
+
+fun Context.areAppNotificationsEnabled() = NotificationManagerCompat.from(this).areNotificationsEnabled()