+++ clevertap-core/src/main/java/com/clevertap/android/sdk/variables/annotations/Variable.java
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package com.clevertap.android.sdk.feat_variable.annotations;
+package com.clevertap.android.sdk.variables.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;