+++ clevertap-core/src/test/java/com/clevertap/android/sdk/variables/CTVariablesTest.kt
@@ -1,4 +1,4 @@
-package com.clevertap.android.sdk.feat_variable
+package com.clevertap.android.sdk.variables
 
 import com.clevertap.android.sdk.feat_variable.utils.CTVariableUtils.getNameComponents
 import org.junit.*