+++ clevertap-core/src/test/java/com/clevertap/android/sdk/variables/utils/CTVariableUtilsTest.kt
@@ -1,6 +1,6 @@
-package com.clevertap.android.sdk.feat_variable.utils
+package com.clevertap.android.sdk.variables.utils
 
-import com.clevertap.android.sdk.feat_variable.utils.CTVariableUtils.mergeHelper
+import com.clevertap.android.sdk.variables.CTVariableUtils
 import com.clevertap.android.shared.test.BaseTestCase
 import org.junit.*
 import org.junit.Assert.*